TEM_Image_Analysis_Automation_Pipeline
A script I wrote to automate the image analysis of atomic columns of TEM images. Later versions also included the ability to generate descriptive statistics on the data collected. Code was also optimized using standard libraries to maximize good segmentation of atoms, though that is dependent on the image quality as well!